+++ src/renderer/opengl/base/gl_pipeline.h
@@ -25,7 +25,7 @@ public:
     GLPipeline(const sp<Recycler>& recycler, uint32_t version, const String& vertexShader, const String& fragmentShader, const ShaderBindings& bindings);
     virtual ~GLPipeline() override;
 
-    virtual uintptr_t id() override;
+    virtual uint64_t id() override;
     virtual void upload(GraphicsContext& graphicsContext) override;
     virtual RecycleFunc recycle() override;
